KNIME 3.0 API changes. Added PortType and DataType extensions in the registry.
